requirements.txt
jupyter
pandas
numpy
scikit-learn
matplotlib
plotly
seaborn
statsmodels
xlrd
neuraloperator
tqdm
pytorch-lightning
click